requirements.txt
discord.py>=2.3.2
aiohttp>=3.9.0
aiosqlite>=0.19.0
python-dotenv>=1.0.0
cryptography>=42.0.0
fastapi>=0.115.0
uvicorn>=0.32.0